feat: Implement the merchant dashboard home screen and refactor bank account inputs to use a custom component.
